requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.3
djangorestframework==3.14.0
filelock==3.12.2
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.28.1
platformdirs==3.8.1
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.23.1
whitenoise==6.4.0
yarl==1.9.2
